article factories moved to separate folder
